회원가입 및 JWT 기반 로그인 기능 구현(SecurityConfig, AuthController 수정)
