+++ src/main/java/com/kidmily/algoga_server/user/presentation/AuthController.java
@@ -1,7 +1,9 @@
 package com.kidmily.algoga_server.user.presentation;
 
+import com.kidmily.algoga_server.global.annotation.swagger.ApiErrorCodeExample;
 import com.kidmily.algoga_server.global.common.api.response.ApiResponse;
 import com.kidmily.algoga_server.user.application.AuthService;
+import com.kidmily.algoga_server.user.exception.UserErrorCode;
 import com.kidmily.algoga_server.user.presentation.request.AuthLoginRequest;
 import com.kidmily.algoga_server.user.presentation.request.AuthSignupRequest;
 import com.kidmily.algoga_server.user.presentation.response.AuthTokenResponse;
@@ -20,11 +22,11 @@ public class AuthController {
     private final AuthService authService;
 
     @Operation(summary = "일반 회원가입")
+    @ApiErrorCodeExample(domain = UserErrorCode.class, value = {"ALREADY_EXISTS_EMAIL"})
     @PostMapping("/signup")
     public ApiResponse<Void> signup(@RequestBody @Valid AuthSignupRequest request) {
         authService.signup(request);
 
-        // ⭐️ 수정됨: 공통 ApiResponse의 created 메서드 사용 (201 Created)
         return ApiResponse.created(
                 "AUTH_SIGNUP_SUCCESS",
                 "회원가입이 완료되었습니다.",
@@ -33,11 +35,11 @@ public class AuthController {
     }
 
     @Operation(summary = "일반 로그인")
+    @ApiErrorCodeExample(domain = UserErrorCode.class, value = {"NOT_FOUND_USER", "DELETED_USER", "ACCOUNT_LOCKED", "INVALID_PASSWORD"})
     @PostMapping("/login")
     public ApiResponse<AuthTokenResponse> login(@RequestBody @Valid AuthLoginRequest request) {
         AuthTokenResponse response = authService.login(request);
 
-        // ⭐️ 수정됨: 공통 ApiResponse의 success 메서드 사용 (200 OK)
         return ApiResponse.success(
                 "AUTH_LOGIN_SUCCESS",
                 "로그인에 성공했습니다.",
